+++ azure/rs_azure_compute/compute_test_cat.rb
@@ -30,6 +30,52 @@ resource "my_availability_group", type: "rs_azure_compute.availability_set" do
   } end
 end
 
+resource "deployment_availability_group", type: "rs_azure_compute.availability_set" do
+  name @@deployment.name
+  resource_group @@deployment.name
+  location "Central US"
+  sku do {
+    "name" => "Aligned"
+  } end
+  properties do {
+      "platformUpdateDomainCount" => 5,
+      "platformFaultDomainCount" => 3
+  } end
+end
+
+resource "server1", type: "server" do
+  name join(["server1-", last(split(@@deployment.href, "/"))])
+  cloud "AzureRM Central US"
+  server_template "RightLink 10.6.0 Linux Base"
+  multi_cloud_image_href "/api/multi_cloud_images/423486003"
+  network "ARM-CentralUS"
+  subnets "default"
+  instance_type "Standard_F1"
+  security_groups "Default"
+  associate_public_ip_address true
+  cloud_specific_attributes do {
+    "availability_set" => @my_availability_group.name
+  }
+  end
+end
+
+resource "my_vm_extension", type: "rs_azure_compute.extensions" do
+  name join(["dokku-", last(split(@@deployment.href, "/"))])
+  resource_group @@deployment.name
+  location "Central US"
+  virtualMachineName @server1.name
+  properties do {
+    "publisher" => "Microsoft.OSTCExtensions",
+    "type" => "CustomScriptForLinux",
+    "typeHandlerVersion" => "1.5",
+    "autoUpgradeMinorVersion" => true,
+    "settings" => {
+       "fileUris" => [ "https://raw.githubusercontent.com/Azure/azure-quickstart-templates/master/dokku-vm/deploy_dokku.sh"],
+       "commandToExecute" => "sh deploy_dokku.sh 0.5.8"
+    }
+  } end
+end
+
 operation "launch" do
  description "Launch the application"
  definition "launch_handler"
@@ -38,10 +84,18 @@ operation "launch" do
   } end
 end
 
-define launch_handler(@my_availability_group) return @my_availability_group,$vms do
+define launch_handler(@my_availability_group,@server1,@my_vm_extension,@deployment_availability_group) return @my_availability_group,@my_availability_group,@server1,@my_vm_extension,@deployment_availability_group,$vms do
   call start_debugging()
   provision(@my_availability_group)
+  provision(@server1)
+  provision(@my_vm_extension)
+  $object = to_object(@deployment_availability_group)
+  $fields = $object["fields"]
+  $name = $fields["name"]
+  $resource_group = $fields["resource_group"]
+  @deployment_ag = rs_azure_compute.availability_set.show(resource_group: $resource_group, name: $name)
   $vms = to_s(@my_availability_group.virtualmachines)
+  $vms = $vms + to_s(@deployment_ag.virtualmachines)
   call sys_log.detail("vms:" + to_s($vms))
   call stop_debugging()
 end