+++ azure/rs_azure_compute/changelog.md
@@ -1,5 +1,10 @@
 Compute Plugin changelog
 
+v1.1
+-----
+- adding compute extensions
+- moving from vm_name to virtualMachineName
+
 v1.0
 -----
 - initial release